Test only for integer expressions to warn about them
Remove the rest of the data types tests because it is not clear
that they are valid tests (i.e. they are supported by any compilers).

Also, move if_not_logical test to sema because it is not a valid
gfortran tests, so we will check only that OFC can parse it and
provides semantic analysis.